refactor: 피드백반영 - 메인함수 별도 분리
